Prevent eels from spawning before depth 5
The most common tactic for dealing with them seemed to be to
retreat to land and wait an indeterminate amount of time for them
to leave. The best place for them is probably after the player
has a more interesting way to deal with them but before the more
threatening aquatic enemies appear. This should also keep
autoexplore viable a bit longer.